+Renamed abstractAdapter -> abstract-adapter
