+++ src/adapters/abstract-adapter.js
